rectangle layouter moved to package
